монета_дробная_ставка_тест_пробовать_бесплатно

+++ rec_file.py
@@ -304,12 +304,12 @@ def moneta_stavka(id,stavka,or_or_re,fileRead):
                 perc=1
                 stavka=stavka[:-1:]
                 try:
-                    int(stavka)
+                    float(stavka)
                 except:
                     return ("Неверное использование процентной ставки. Использование: <1%-100%> <орел/решка>")  
-                if (len(stavka)>4 or int(stavka)>100 or int(stavka)<1):
+                if (len(stavka)>4 or float(stavka)>100 or float(stavka)<1):
                     return ("Неверное использование процентной ставки. Процент должен быть от 1 до 100")
-                stavka=get_balance(id, fileRead)*int(stavka)//100
+                stavka=get_balance(id, fileRead)*float(stavka)//100
         if (perc==0):
             try:
                 int(stavka)
